+++ docs/metro/js/metro.js
@@ -119,7 +119,7 @@ var isTouch = (('ontouchstart' in window) || (navigator.MaxTouchPoints > 0) || (
 var Metro = {
 
     version: "4.2.42",
-    compileTime: "07/05/2019 20:25:28",
+    compileTime: "07/05/2019 20:40:14",
     buildNumber: "723",
     isTouchable: isTouch,
     fullScreenEnabled: document.fullscreenEnabled,
@@ -20105,13 +20105,13 @@ var Streamer = {
         this.changeSource();
     },
 
-    data: function(s){
+    dataSet: function(s){
         if (s === undefined) {
-            return this.options.source;
+            return this.options.data;
         }
 
         this.options.data = s;
-        this.changeData();
+        this.changeData(s);
     },
 
     getStreamerData: function(){
@@ -20194,7 +20194,10 @@ var Streamer = {
         var that = this, element = this.element, o = this.options;
         var old_data = this.data;
 
-        o.data =  data ? data : JSON.parse(element.attr("data-data"));
+        console.log(typeof data);
+        console.log(data);
+
+        o.data =  typeof data === 'object' ? data : JSON.parse(element.attr("data-data"));
 
         this.data = o.data;